Atualizações API v2

+++ ws/app/libs/Log.php
@@ -31,7 +31,6 @@ class Log {
      * Grava dados no log
      */
     private static function setLog($tipo, $log) {
-        $file = __DIR__ . "/../logs/" . DOMINIO . "_ws_" . date('Y-m-d') . ".txt";
         if (is_array($log)) {
             $str = date('d/m/Y H:i:s') . " {$tipo}: ";
             $str .= print_r($log, true);
@@ -46,6 +45,30 @@ class Log {
             $str = date('d/m/Y H:i:s') . " {$tipo}: {$log}\n";
         }
 
+        $logdir = __DIR__ . "/../logs/debug/";
+        $file = $logdir . DOMINIO . ".txt";
+        if (file_exists($file)) {
+            $trace = [];
+            foreach (debug_backtrace() as $k => $v) {
+                if ($k == 0) {
+                    continue;
+                }
+                if (array_key_exists('file', $v)) {
+                    foreach ($v['args'] as &$arg) {
+                        if (is_object($arg)) {
+                            $arg = '(Object)';
+                        }
+                    }
+                    array_push($trace, array_filter($v, function($key) {
+                                return $key != 'object';
+                            }, ARRAY_FILTER_USE_KEY));
+                }
+            }
+            $str .= "Trace: " . print_r($trace, true);
+        } else {
+            $logdir = __DIR__ . "/../logs/";
+            $file = __DIR__ . "/../logs/" . DOMINIO . "_ws_" . date('Y-m-d') . ".txt";
+        }
         $exists = file_exists($file);
         file_put_contents($file, $str, FILE_APPEND | FILE_TEXT);
         //Se criou o arquivo agora
@@ -54,10 +77,8 @@ class Log {
             @chown($file, "apache");
             @chgrp($file, "brudam");
             //Apagar arquivos com mais de 5 dias
-            foreach (new DirectoryIterator(__DIR__ . "/../logs") as $fileInfo) {
-                if ($fileInfo->isDot())
-                    continue;
-                if (time() - $fileInfo->getCTime() >= 432000) {//5 dias ( 5 * 24 * 60 * 60)
+            foreach (new DirectoryIterator($logdir) as $fileInfo) {
+                if (!$fileInfo->isDot() && !$fileInfo->isDir() && time() - $fileInfo->getCTime() >= 432000) {//5 dias ( 5 * 24 * 60 * 60)
                     unlink($fileInfo->getRealPath());
                 }
             }
